Add files via upload
The main idea of the project is to create a website where visitors can read and write game reviews, since the most of comments are untrustworthy, the players wary about the credibility of comments on games, and some of them get disappointed because the game didn`t meet their expectations, this is because often there is no clear vision and regarding the games. On the other hand, parents confused in choosing the appropriate games for their children due to their lack of knowledge in this field and the lack of reference in it.
So, to solve all these problems, the site will provide a clear vision, and honest references so it assists the players in learning more about the game and it will enhance the players' experience and make it easier for them and the parents to make decisions.
Through our site, we aim to enable parents to manage the games their children play through comments, make developers more familiar with the advantages and disadvantages of applications, help the streamer decide which game viewers prefer to improve the quality of their content, and help 
investors decide most successful games, feedback can also help the customer determine if the game is reliable and developed by a trusted developer.
The main contents of this document are divided into five sections with the introduction, which is the first section, we have section 2 is domain analysis which content the terminology, general domain knowledge, customers and users, the environment, tasks and procedures, competing software, similarities across domains and organizations. In section 3 the requirements engineering you will find the system users, use case diagram and product backlog. Also, in section 4 we will talk about the system architecture and finally we have the references.